# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Music Rhythm Analysis API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AnalyzeMusicBody {
  /** URL of the song to analyze (YouTube, SoundCloud, or direct audio link) */
  url: string;
}

export interface AnalyzePlaylistBody {
  /**
   * Array of song URLs to analyze
   * @minItems 1
   * @maxItems 20
   */
  urls: string[];
}

export type PlaylistTrackResultStatus =
  (typeof PlaylistTrackResultStatus)[keyof typeof PlaylistTrackResultStatus];

export const PlaylistTrackResultStatus = {
  success: "success",
  error: "error",
} as const;

export interface FrequencyBand {
  /** Band label (e.g. Sub Bass, Bass, Low Mid, Mid, Upper Mid, Presence, Brilliance) */
  label: string;
  minHz: number;
  maxHz: number;
  /** Average amplitude 0-1 */
  amplitude: number;
}

export interface TempoChange {
  /** Time in seconds */
  timestamp: number;
  bpm: number;
}

export interface HealingFrequency {
  /** Frequency in Hz */
  frequency: number;
  /** Name of the healing frequency (e.g. Solfeggio 528 Hz) */
  name: string;
  /** Presence strength 0-1 */
  presence: number;
  /** Associated health benefit */
  benefit: string;
}

export type CellularResonanceCategory =
  (typeof CellularResonanceCategory)[keyof typeof CellularResonanceCategory];

export const CellularResonanceCategory = {
  highly_beneficial: "highly_beneficial",
  beneficial: "beneficial",
  neutral: "neutral",
  potentially_harmful: "potentially_harmful",
} as const;

export interface CellularResonance {
  /** Overall cellular resonance score 0-100 */
  score: number;
  healingFrequencies: HealingFrequency[];
  /** Human-readable assessment of cellular impact */
  assessment: string;
  category: CellularResonanceCategory;
}

export interface MoodDimensions {
  /** Energy dimension 0-1 */
  energy: number;
  /** Aggression dimension 0-1 */
  aggression: number;
  /** Euphoria dimension 0-1 */
  euphoria: number;
  /** Tension dimension 0-1 */
  tension: number;
  /** Calmness dimension 0-1 */
  calmness: number;
}

export interface MoodGenre {
  /** Primary detected mood (e.g. Energetic, Calm, Aggressive, Euphoric) */
  mood: string;
  /** Emoji representing the mood */
  moodEmoji: string;
  /** Mood detection confidence 0-1 */
  moodConfidence: number;
  /** Human-readable description of the detected mood */
  moodDescription: string;
  /** Suggested primary genre (e.g. EDM, Jazz, Lo-Fi) */
  genre: string;
  /** Suggested sub-genre */
  subGenre: string;
  /** Genre detection confidence 0-1 */
  genreConfidence: number;
  /** Key sonic characteristics of the track */
  characteristics: string[];
  moodDimensions: MoodDimensions;
}

export interface MusicAnalysis {
  id: number;
  url: string;
  title: string;
  /** Beats per minute */
  bpm: number;
  /** Musical key (e.g. C major, A minor) */
  key: string;
  /** Energy level 0-1 */
  energy: number;
  /** Danceability score 0-1 */
  danceability: number;
  /** Dominant frequency in Hz */
  dominantFrequency: number;
  frequencySpectrum: FrequencyBand[];
  /** Beat intervals in seconds */
  beatPattern: number[];
  tempoChanges: TempoChange[];
  cellularResonance: CellularResonance;
  moodGenre?: MoodGenre;
  createdAt: string;
}

export interface PlaylistTrackResult {
  url: string;
  status: PlaylistTrackResultStatus;
  analysis?: MusicAnalysis;
  error?: string;
}

export interface PlaylistAnalysisResult {
  total: number;
  succeeded: number;
  failed: number;
  results: PlaylistTrackResult[];
}

export type MusicAnalysisSummaryCategory =
  (typeof MusicAnalysisSummaryCategory)[keyof typeof MusicAnalysisSummaryCategory];

export const MusicAnalysisSummaryCategory = {
  highly_beneficial: "highly_beneficial",
  beneficial: "beneficial",
  neutral: "neutral",
  potentially_harmful: "potentially_harmful",
} as const;

export interface MusicAnalysisSummary {
  id: number;
  title: string;
  url: string;
  bpm: number;
  key: string;
  energy: number;
  cellularScore: number;
  category: MusicAnalysisSummaryCategory;
  createdAt: string;
}

export type AnalysisStatsCategoryBreakdown = {
  highly_beneficial: number;
  beneficial: number;
  neutral: number;
  potentially_harmful: number;
};

export interface AnalysisStats {
  totalAnalyses: number;
  averageBpm: number;
  averageCellularScore: number;
  topHealingFrequency: string;
  categoryBreakdown: AnalysisStatsCategoryBreakdown;
}

export interface LyricsResult {
  /** Full lyrics text or null if not found */
  lyrics?: string | null;
  /** Parsed or provided artist name */
  artist: string;
  /** Parsed or provided song title */
  song: string;
  /** Whether lyrics were successfully found */
  found: boolean;
  /** Lyrics split into individual lines */
  lines?: string[];
  /** Additional info message (e.g. if not found) */
  message?: string;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export type GetRecentAnalysesParams = {
  /**
   * @minimum 1
   * @maximum 50
   */
  limit?: number;
};

export type GetLyricsParams = {
  /**
   * Full song title (e.g. "Artist - Song Name")
   */
  title: string;
  /**
   * Explicit artist name override
   */
  artist?: string;
  /**
   * Explicit song name override
   */
  song?: string;
};