# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Music Rhythm Analysis API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: analysis
    description: Music analysis operations
  - name: lyrics
    description: Song lyrics operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /analyze:
    post:
      operationId: analyzeMusic
      tags: [analysis]
      summary: Analyze music from a URL
      description: Accepts a song URL and returns rhythmic analysis data including BPM, beat patterns, frequency spectrum, and cellular resonance assessment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnalyzeMusicBody"
      responses:
        "200":
          description: Analysis results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MusicAnalysis"
        "400":
          description: Invalid URL or unsupported format
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Analysis failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /analyze-playlist:
    post:
      operationId: analyzePlaylist
      tags: [analysis]
      summary: Analyze multiple songs from URLs
      description: Accepts multiple song URLs and returns rhythmic analysis for each
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnalyzePlaylistBody"
      responses:
        "200":
          description: Playlist analysis results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlaylistAnalysisResult"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /analyses:
    get:
      operationId: getRecentAnalyses
      tags: [analysis]
      summary: Get recent analyses
      description: Returns a list of recent music analyses
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 50
      responses:
        "200":
          description: List of recent analyses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MusicAnalysisSummary"
  /analyses/{id}:
    get:
      operationId: getAnalysis
      tags: [analysis]
      summary: Get a specific analysis by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Analysis details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MusicAnalysis"
        "404":
          description: Analysis not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /analyses/stats:
    get:
      operationId: getAnalysisStats
      tags: [analysis]
      summary: Get aggregated analysis statistics
      description: Returns summary statistics across all analyses
      responses:
        "200":
          description: Analysis statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalysisStats"
  /lyrics:
    get:
      operationId: getLyrics
      tags: [lyrics]
      summary: Get lyrics for a song
      description: Fetches lyrics for a song by parsing the video title into artist and song name
      parameters:
        - name: title
          in: query
          required: true
          schema:
            type: string
          description: Full song title (e.g. "Artist - Song Name")
        - name: artist
          in: query
          required: false
          schema:
            type: string
          description: Explicit artist name override
        - name: song
          in: query
          required: false
          schema:
            type: string
          description: Explicit song name override
      responses:
        "200":
          description: Lyrics result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LyricsResult"
        "404":
          description: Lyrics not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    AnalyzeMusicBody:
      type: object
      properties:
        url:
          type: string
          description: URL of the song to analyze (YouTube, SoundCloud, or direct audio link)
      required:
        - url
    AnalyzePlaylistBody:
      type: object
      properties:
        urls:
          type: array
          items:
            type: string
          description: Array of song URLs to analyze
          minItems: 1
          maxItems: 20
      required:
        - urls
    PlaylistAnalysisResult:
      type: object
      properties:
        total:
          type: integer
        succeeded:
          type: integer
        failed:
          type: integer
        results:
          type: array
          items:
            $ref: "#/components/schemas/PlaylistTrackResult"
      required:
        - total
        - succeeded
        - failed
        - results
    PlaylistTrackResult:
      type: object
      properties:
        url:
          type: string
        status:
          type: string
          enum: [success, error]
        analysis:
          $ref: "#/components/schemas/MusicAnalysis"
        error:
          type: string
      required:
        - url
        - status
    MusicAnalysis:
      type: object
      properties:
        id:
          type: integer
        url:
          type: string
        title:
          type: string
        bpm:
          type: number
          description: Beats per minute
        key:
          type: string
          description: Musical key (e.g. C major, A minor)
        energy:
          type: number
          description: Energy level 0-1
        danceability:
          type: number
          description: Danceability score 0-1
        dominantFrequency:
          type: number
          description: Dominant frequency in Hz
        frequencySpectrum:
          type: array
          items:
            $ref: "#/components/schemas/FrequencyBand"
        beatPattern:
          type: array
          items:
            type: number
          description: Beat intervals in seconds
        tempoChanges:
          type: array
          items:
            $ref: "#/components/schemas/TempoChange"
        cellularResonance:
          $ref: "#/components/schemas/CellularResonance"
        moodGenre:
          $ref: "#/components/schemas/MoodGenre"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - url
        - title
        - bpm
        - key
        - energy
        - danceability
        - dominantFrequency
        - frequencySpectrum
        - beatPattern
        - tempoChanges
        - cellularResonance
        - createdAt
    FrequencyBand:
      type: object
      properties:
        label:
          type: string
          description: Band label (e.g. Sub Bass, Bass, Low Mid, Mid, Upper Mid, Presence, Brilliance)
        minHz:
          type: number
        maxHz:
          type: number
        amplitude:
          type: number
          description: Average amplitude 0-1
      required:
        - label
        - minHz
        - maxHz
        - amplitude
    TempoChange:
      type: object
      properties:
        timestamp:
          type: number
          description: Time in seconds
        bpm:
          type: number
      required:
        - timestamp
        - bpm
    CellularResonance:
      type: object
      properties:
        score:
          type: number
          description: Overall cellular resonance score 0-100
        healingFrequencies:
          type: array
          items:
            $ref: "#/components/schemas/HealingFrequency"
        assessment:
          type: string
          description: Human-readable assessment of cellular impact
        category:
          type: string
          enum: [highly_beneficial, beneficial, neutral, potentially_harmful]
      required:
        - score
        - healingFrequencies
        - assessment
        - category
    HealingFrequency:
      type: object
      properties:
        frequency:
          type: number
          description: Frequency in Hz
        name:
          type: string
          description: Name of the healing frequency (e.g. Solfeggio 528 Hz)
        presence:
          type: number
          description: Presence strength 0-1
        benefit:
          type: string
          description: Associated health benefit
      required:
        - frequency
        - name
        - presence
        - benefit
    MoodDimensions:
      type: object
      properties:
        energy:
          type: number
          description: Energy dimension 0-1
        aggression:
          type: number
          description: Aggression dimension 0-1
        euphoria:
          type: number
          description: Euphoria dimension 0-1
        tension:
          type: number
          description: Tension dimension 0-1
        calmness:
          type: number
          description: Calmness dimension 0-1
      required:
        - energy
        - aggression
        - euphoria
        - tension
        - calmness
    MoodGenre:
      type: object
      properties:
        mood:
          type: string
          description: Primary detected mood (e.g. Energetic, Calm, Aggressive, Euphoric)
        moodEmoji:
          type: string
          description: Emoji representing the mood
        moodConfidence:
          type: number
          description: Mood detection confidence 0-1
        moodDescription:
          type: string
          description: Human-readable description of the detected mood
        genre:
          type: string
          description: Suggested primary genre (e.g. EDM, Jazz, Lo-Fi)
        subGenre:
          type: string
          description: Suggested sub-genre
        genreConfidence:
          type: number
          description: Genre detection confidence 0-1
        characteristics:
          type: array
          items:
            type: string
          description: Key sonic characteristics of the track
        moodDimensions:
          $ref: "#/components/schemas/MoodDimensions"
      required:
        - mood
        - moodEmoji
        - moodConfidence
        - moodDescription
        - genre
        - subGenre
        - genreConfidence
        - characteristics
        - moodDimensions
    MusicAnalysisSummary:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        url:
          type: string
        bpm:
          type: number
        key:
          type: string
        energy:
          type: number
        cellularScore:
          type: number
        category:
          type: string
          enum: [highly_beneficial, beneficial, neutral, potentially_harmful]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - url
        - bpm
        - key
        - energy
        - cellularScore
        - category
        - createdAt
    AnalysisStats:
      type: object
      properties:
        totalAnalyses:
          type: integer
        averageBpm:
          type: number
        averageCellularScore:
          type: number
        topHealingFrequency:
          type: string
        categoryBreakdown:
          type: object
          properties:
            highly_beneficial:
              type: integer
            beneficial:
              type: integer
            neutral:
              type: integer
            potentially_harmful:
              type: integer
          required:
            - highly_beneficial
            - beneficial
            - neutral
            - potentially_harmful
      required:
        - totalAnalyses
        - averageBpm
        - averageCellularScore
        - topHealingFrequency
        - categoryBreakdown
    LyricsResult:
      type: object
      properties:
        lyrics:
          type: string
          nullable: true
          description: Full lyrics text or null if not found
        artist:
          type: string
          description: Parsed or provided artist name
        song:
          type: string
          description: Parsed or provided song title
        found:
          type: boolean
          description: Whether lyrics were successfully found
        lines:
          type: array
          items:
            type: string
          description: Lyrics split into individual lines
        message:
          type: string
          description: Additional info message (e.g. if not found)
      required:
        - found
        - artist
        - song
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message