# lib/api-zod/src/generated/types/moodDimensions.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Music Rhythm Analysis API
 * OpenAPI spec version: 0.1.0
 */

export interface MoodDimensions {
  /** Energy dimension 0-1 */
  energy: number;
  /** Aggression dimension 0-1 */
  aggression: number;
  /** Euphoria dimension 0-1 */
  euphoria: number;
  /** Tension dimension 0-1 */
  tension: number;
  /** Calmness dimension 0-1 */
  calmness: number;
}