# lib/api-zod/src/generated/types/moodGenre.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Music Rhythm Analysis API
 * OpenAPI spec version: 0.1.0
 */
import type { MoodDimensions } from "./moodDimensions";

export interface MoodGenre {
  /** Primary detected mood (e.g. Energetic, Calm, Aggressive, Euphoric) */
  mood: string;
  /** Emoji representing the mood */
  moodEmoji: string;
  /** Mood detection confidence 0-1 */
  moodConfidence: number;
  /** Human-readable description of the detected mood */
  moodDescription: string;
  /** Suggested primary genre (e.g. EDM, Jazz, Lo-Fi) */
  genre: string;
  /** Suggested sub-genre */
  subGenre: string;
  /** Genre detection confidence 0-1 */
  genreConfidence: number;
  /** Key sonic characteristics of the track */
  characteristics: string[];
  moodDimensions: MoodDimensions;
}