# lib/api-zod/src/generated/types/musicAnalysis.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Music Rhythm Analysis API
 * OpenAPI spec version: 0.1.0
 */
import type { CellularResonance } from "./cellularResonance";
import type { FrequencyBand } from "./frequencyBand";
import type { MoodGenre } from "./moodGenre";
import type { TempoChange } from "./tempoChange";

export interface MusicAnalysis {
  id: number;
  url: string;
  title: string;
  /** Beats per minute */
  bpm: number;
  /** Musical key (e.g. C major, A minor) */
  key: string;
  /** Energy level 0-1 */
  energy: number;
  /** Danceability score 0-1 */
  danceability: number;
  /** Dominant frequency in Hz */
  dominantFrequency: number;
  frequencySpectrum: FrequencyBand[];
  /** Beat intervals in seconds */
  beatPattern: number[];
  tempoChanges: TempoChange[];
  cellularResonance: CellularResonance;
  moodGenre?: MoodGenre;
  createdAt: Date;
}